code/requirements.txt
# Основные зависимости для CI/CD системы тестирования 1С
# Основан на стандартах тестирования из docs/1c_testing_standards.md

# ===============================
# Инструменты тестирования
# ===============================

# Основной фреймворк тестирования
pytest==7.4.3
pytest-xdist==3.3.1
pytest-html==3.2.0
pytest-cov==4.1.0
pytest-json-report==1.5.0
pytest-timeout==2.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1

# Специализированные плагины для 1С
xunit-for-1c-client==2.1.0
vanessa-automation==2.2.5

# BDD тестирование
behave==1.2.6
allure-behave==2.13.2
allure-pytest==2.13.2

# HTTP тестирование
httpx==0.25.1
aiohttp==3.8.6
requests==2.31.0
responses==0.23.3

# Browser тестирование
selenium==4.14.0
playwright==1.40.0

# ===============================
# Инструменты качества кода
# ===============================

# Линтинг и форматирование
flake8==6.1.0
black==23.9.1
isort==5.12.0
mypy==1.6.1
pylint==3.0.2
pylint-django==2.5.5

# Статический анализ
radon==6.0.1
vulture==2.7
xenon==0.9.1
bandit==1.7.5
safety==2.3.4
semgrep==1.38.0

# Анализ сложности
mccabe==0.7.0
cyclomatic==0.6.0

# ===============================
# Покрытие кода и метрики
# ===============================

# Основные инструменты покрытия
coverage==7.3.2
pytest-cov==4.1.0

# Интеграция с системами отчетности
codacy-coverage==3.1.0
codecov==2.1.13
codeclimate-test-reporter==0.2.0

# Отчетность и документация
allure-python-commons==2.13.2
reportlab==4.0.7
matplotlib==3.7.2
pandas==2.1.3
numpy==1.24.3

# ===============================
# Управление данными и фикстурами
# ===============================

# Генерация тестовых данных
faker==19.12.0
factory-boy==3.3.0
hypothesis==6.90.0

# Работа с фикстурами
freezegun==1.2.2
responses==0.23.3
requests-mock==1.11.0

# ===============================
# Базы данных и миграции
# ===============================

# Основные драйверы БД
psycopg2-binary==2.9.9
pymysql==1.1.0
pyodbc==4.0.39
sqlalchemy==2.0.23
alembic==1.12.1

# ORM и дополнительные возможности
django-orm==4.2.7
sqlalchemy-utils==0.41.1

# ===============================
# HTTP клиенты и API
# ===============================

# Современные HTTP клиенты
httpx==0.25.1
aiohttp==3.8.6
requests==2.31.0

# API фреймворки
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# API документация
swagger-ui-bundle==0.0.9
redoc==2.1.0

# ===============================
# Логирование и мониторинг
# ===============================

# Логирование
loguru==0.7.2
structlog==23.2.0
coloredlogs==15.0.1

# Мониторинг и метрики
prometheus-client==0.19.0
psutil==5.9.6
statsd==4.0.1

# Трассировка
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0

# ===============================
# Утилиты и вспомогательные библиотеки
# ===============================

# Парсинг данных
lxml==4.9.3
xmltodict==0.13.0
jsonschema==4.19.1
pyyaml==6.0.1
jinja2==3.1.2
toml==0.10.2

# Работа с датами и временем
python-dateutil==2.8.2
pytz==2023.3
arrow==1.3.0

# Работа с конфигурациями
python-dotenv==1.0.0
dynaconf==3.2.4
click==8.1.7

# Работа с файлами и архивами
pathlib2==2.3.7
unidecode==1.3.7
chardet==5.2.0

# ===============================
# Параллелизм и конкурентность
# ===============================

# Параллельное выполнение
joblib==1.3.2
concurrent-futures==3.1.1
asyncio-throttle==1.0.2

# Очереди задач
celery==5.3.4
redis==5.0.1

# ===============================
# Интеграция с системами CI/CD
# ===============================

# Jenkins интеграция
python-jenkins==1.8.0
jenkinsapi==1.7.0

# GitLab интеграция
python-gitlab==4.0.0

# GitHub интеграция
PyGithub==1.59.1

# SonarQube интеграция
sonarqube-api==2.7.0
sonar-scanner==5.0.0

# ===============================
# Нагрузочное тестирование
# ===============================

# Основные инструменты нагрузки
locust==2.16.1
requests-mock==1.11.0

# Профилирование производительности
memory-profiler==0.61.0
line-profiler==4.1.1
py-spy==0.3.14

# Визуализация производительности
snakeviz==2.1.1
gprof2dot==2022.7.29

# ===============================
# Интеграционное тестирование
# ===============================

# Контейнеризация для тестов
testcontainers==3.7.1
docker==6.1.3

# Моки и заглушки
responses==0.23.3
httpretty==1.1.4
mock==5.1.0

# ===============================
# Безопасность и сканирование
# ===============================

# Сканирование уязвимостей
bandit==1.7.5
safety==2.3.4
semgrep==1.38.0
snyk==1.1207.0

# Шифрование и хеширование
cryptography==41.0.7
hashlib2==20081119

# ===============================
# Работа с 1С специфичными форматами
# ===============================

# Парсинг файлов 1С
epflib==2.2.0
erflib==1.8.0
v8unpack==0.7.1

# Работа с конфигурациями 1С
cf-parser==1.4.2
confupdate==1.2.1

# ===============================
# Вспомогательные библиотеки
# ===============================

# Обработка ошибок и исключений
sentry-sdk==1.38.0
rollbar==0.16.3

# Кэширование
cachetools==5.3.2
redis==5.0.1

# Уведомления
slack-sdk==3.23.0
telegram-send==0.33

# Работа с изображениями и отчетами
Pillow==10.1.0
weasyprint==60.2

# ===============================
# Разработческие зависимости
# ===============================

# Интерактивная разработка
ipython==8.16.1
jupyter==1.0.0
ipdb==0.13.13

# Отладка и профилирование
debugpy==1.8.0
pudb==2023.1

# Красивый вывод в консоль
rich==13.6.0
typer==0.9.0
colorama==0.4.6

# ===============================
# Версионирование и документация
# ===============================

# Генерация документации
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
myst-parser==2.0.0

# Версионирование
bump2version==1.0.1
semantic-version==2.10.0

# ===============================
# Минимальные версии Python
# ===============================

# Требования к версиям Python
# Python >= 3.8 (рекомендуется 3.11 для лучшей производительности)

# Основные зависимости с минимальными версиями
# pytest >= 7.0.0
# coverage >= 7.0.0
# requests >= 2.28.0
# pyyaml >= 6.0.0

# ===============================
# Опциональные зависимости
# ===============================

# Для Windows (опционально)
# pywin32==306; sys_platform == "win32"

# Для Linux системных интеграций (опционально)
# python-systemd==235; sys_platform == "linux"

# Для macOS (опционально)
# pyobjc==10.0; sys_platform == "darwin"